pyproject.toml
[tool.poetry]
name = "hac3"
version = "0.1.0"
description = ""
authors = ["Bekbolot001 <bekbolotboronchiev.bt.966@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.9"
python-decouple = "^3.8"
pillow = "^10.3.0"
django-filter = "^24.2"
django-jazzmin = "^3.0.0"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
